Reworked split up of files
