+++ src/__init__.py
@@ -18,5 +18,6 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-from .main import Workbook, excel2df, Sheet, Range
+from .main import Workbook, excel2df, Sheet
+from .range import Range
 from .tools import number_to_date, date_to_number
